+++ src/components/import/ImportPage.jsx
@@ -10,10 +10,10 @@ import {
     getPeriods,
     periodTypes,
 } from '../../utils/time.js'
+import Dataset from '../shared/Dataset.jsx'
 import GEETokenCheck from '../shared/GEETokenCheck.jsx'
 import Resolution from '../shared/Resolution.jsx'
 import DataElement from './DataElement.jsx'
-import Dataset from './Dataset.jsx'
 import ExtractData from './ExtractData.jsx'
 import OrgUnits from './OrgUnits.jsx'
 import Period from './Period.jsx'
@@ -78,8 +78,7 @@ const ImportPage = () => {
                         <Period
                             calendar={calendar}
                             period={period}
-                            datasetPeriodType={dataset?.periodType}
-                            datasetPeriod={dataset?.period}
+                            dataset={dataset}
                             onChange={setPeriod}
                         />
                         <OrgUnits selected={orgUnits} onChange={setOrgUnits} />