+++ src/components/import/Period.jsx
@@ -3,10 +3,11 @@ import i18n from '@dhis2/d2-i18n'
 import { CalendarInput } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import { useEffect } from 'react'
-import { DAILY, SIXTEEN_DAYS, WEEKLY } from '../../utils/time.js'
+import { DAILY, WEEKLY, SIXTEEN_DAYS, YEARLY } from '../../utils/time.js'
 import TimeZone from '../shared/TimeZone.jsx'
 import PeriodType from './PeriodType.jsx'
 import styles from './styles/Period.module.css'
+import YearRange from './YearRange.jsx'
 
 const userSettingsQuery = {
     userSettings: {
@@ -17,17 +18,18 @@ const userSettingsQuery = {
     },
 }
 
-const Period = ({
-    calendar,
-    period,
-    datasetPeriodType,
-    datasetPeriod,
-    onChange,
-}) => {
+const Period = ({ calendar, period, dataset = {}, onChange }) => {
     const result = useDataQuery(userSettingsQuery)
     const { data: { userSettings: { keyUiLocale: locale } = {} } = {} } = result
+    const {
+        periodType: datasetPeriodType,
+        period: datasetPeriod,
+        minYear,
+        maxYear,
+    } = dataset
     const { periodType, startTime, endTime } = period
     const hasNoPeriod = datasetPeriodType === 'N/A'
+    const isYearly = datasetPeriodType === YEARLY
 
     // Set period locale from user settings
     useEffect(() => {
@@ -37,7 +39,10 @@ const Period = ({
     }, [locale, onChange, period])
 
     useEffect(() => {
-        if (datasetPeriodType === SIXTEEN_DAYS && period.periodType === DAILY) {
+        if (
+            datasetPeriodType === SIXTEEN_DAYS &&
+            [DAILY, YEARLY].includes(period.periodType)
+        ) {
             onChange({ ...period, periodType: WEEKLY })
         }
     }, [period, datasetPeriodType, onChange])
@@ -45,14 +50,23 @@ const Period = ({
     return (
         <div className={styles.container}>
             <h2>{i18n.t('Period')}</h2>
-            {hasNoPeriod ? (
+            {hasNoPeriod && (
                 <p>
                     {i18n.t(
                         'The data will be assigned a default yearly period that matches the year it was collected: {{datasetPeriod}}',
                         { datasetPeriod, nsSeparator: ';' }
                     )}
                 </p>
-            ) : (
+            )}
+            {isYearly && (
+                <YearRange
+                    period={period}
+                    minYear={minYear}
+                    maxYear={maxYear}
+                    onChange={onChange}
+                />
+            )}
+            {!hasNoPeriod && !isYearly && (
                 <>
                     <p>
                         {i18n.t(
@@ -104,8 +118,7 @@ const Period = ({
 Period.propTypes = {
     onChange: PropTypes.func.isRequired,
     calendar: PropTypes.string,
-    datasetPeriod: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-    datasetPeriodType: PropTypes.string,
+    dataset: PropTypes.object,
     period: PropTypes.object,
 }